+++ stress-tests/graph/MaximumClique.cpp
@@ -31,7 +31,7 @@ int main() {
         }
         Maxclique clique2(ed);
         ll mx = 0;
-        maximal::cliques(ed2, [&](auto x){mx = max(mx, int(x.count()));});
+        maximal::cliques(ed2, [&](auto x){mx = max(mx, ll(x.count()));});
         assert(mx == sz(clique2.maxClique()));
     }
     cout<<"Tests passed!"<<endl;
