+++ stress-tests/utilities/randGeo.h
@@ -4,5 +4,5 @@
 
 template <class T=double>
 Point<T> randIntPt(ll lim) {
-    return Point<T>(rand()%(lim*2) - lim, rand()%(lim*2)-lim);
+    return Point<T>((double)(rand()%(lim*2) - lim), (double)(rand()%(lim*2)-lim));
 }
